Brew install

+++ brew.sh
@@ -1,13 +1,16 @@
 #!/usr/bin/env bash
 
-# Install command-line tools using Homebrew.
+# Install command-line tools and apps using Homebrew.
+brew bundle
 
 # Save Homebrew’s installed location.
 BREW_PREFIX=$(brew --prefix)
 
 # Install GNU core utilities (those that come with macOS are outdated).
 # Don’t forget to add `$(brew --prefix coreutils)/libexec/gnubin` to `$PATH`.
-ln -s "${BREW_PREFIX}/bin/gsha256sum" "${BREW_PREFIX}/bin/sha256sum"
+if [ ! -f "${BREW_PREFIX}/bin/sha256sum" ]; then
+  ln -s "${BREW_PREFIX}/bin/gsha256sum" "${BREW_PREFIX}/bin/sha256sum";
+fi;
 
 # Install Bash 4.
 # Switch to using brew-installed bash as default shell
